src/video-broker/requirements.txt
python-qpid-proton
numpy
